+++ packages/synapse-interface/components/Portfolio/Transaction/MostRecentTransaction.tsx
@@ -123,7 +123,7 @@ export const MostRecentTransaction = () => {
     if (!seenHistoricalTransactions || !userHistoricalTransactions) {
       return false
     } else {
-      return seenHistoricalTransactions.some(
+      return seenHistoricalTransactions?.some(
         (transaction: BridgeTransaction) =>
           transaction.kappa ===
           (lastHistoricalTransaction.kappa as BridgeTransaction)
@@ -160,8 +160,8 @@ export const MostRecentTransaction = () => {
             startedTimestamp={
               transaction.timestamp ? transaction.timestamp : transaction.id
             }
-            transactionHash={transaction.transactionHash as string}
-            isSubmitted={transaction.isSubmitted as boolean}
+            transactionHash={transaction.transactionHash}
+            isSubmitted={transaction.isSubmitted}
             transactionType={TransactionType.PENDING}
           />
         </div>
@@ -174,10 +174,10 @@ export const MostRecentTransaction = () => {
         <div data-test-id="most-recent-transaction-pending" className="mt-6">
           <PendingTransaction
             connectedAddress={address as Address}
-            startedTimestamp={transaction?.fromInfo?.time as number}
-            transactionHash={transaction?.fromInfo?.txnHash as string}
+            startedTimestamp={transaction?.fromInfo?.time}
+            transactionHash={transaction?.fromInfo?.txnHash}
             transactionType={TransactionType.PENDING}
-            originValue={transaction?.fromInfo?.value as number}
+            originValue={transaction?.fromInfo?.value}
             originChain={CHAINS_BY_ID[transaction?.fromInfo?.chainID] as Chain}
             destinationChain={
               CHAINS_BY_ID[transaction?.fromInfo?.destinationChainID] as Chain
@@ -217,13 +217,13 @@ export const MostRecentTransaction = () => {
           <PendingTransaction
             connectedAddress={address as Address}
             destinationAddress={transaction?.fromInfo?.address as Address}
-            startedTimestamp={transaction?.fromInfo?.time as number}
-            completedTimestamp={transaction?.toInfo?.time as number}
-            transactionHash={transaction?.fromInfo?.txnHash as string}
-            kappa={transaction?.kappa as string}
+            startedTimestamp={transaction?.fromInfo?.time}
+            completedTimestamp={transaction?.toInfo?.time}
+            transactionHash={transaction?.fromInfo?.txnHash}
+            kappa={transaction?.kappa}
             transactionType={TransactionType.PENDING}
-            originValue={transaction?.fromInfo?.value as number}
-            destinationValue={transaction?.toInfo?.value as number}
+            originValue={transaction?.fromInfo?.value}
+            destinationValue={transaction?.toInfo?.value}
             originChain={CHAINS_BY_ID[transaction?.fromInfo?.chainID] as Chain}
             destinationChain={
               CHAINS_BY_ID[transaction?.fromInfo?.destinationChainID] as Chain