+++ packages/synapse-interface/CHANGELOG.md
@@ -3,6 +3,62 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.183](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.1.182...@synapsecns/synapse-interface@0.1.183) (2023-11-20)
+
+**Note:** Version bump only for package @synapsecns/synapse-interface
+
+
+
+
+
+## [0.1.182](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.1.181...@synapsecns/synapse-interface@0.1.182) (2023-11-15)
+
+**Note:** Version bump only for package @synapsecns/synapse-interface
+
+
+
+
+
+## [0.1.181](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.1.180...@synapsecns/synapse-interface@0.1.181) (2023-11-11)
+
+**Note:** Version bump only for package @synapsecns/synapse-interface
+
+
+
+
+
+## [0.1.180](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.1.179...@synapsecns/synapse-interface@0.1.180) (2023-11-10)
+
+**Note:** Version bump only for package @synapsecns/synapse-interface
+
+
+
+
+
+## [0.1.179](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.1.178...@synapsecns/synapse-interface@0.1.179) (2023-11-09)
+
+**Note:** Version bump only for package @synapsecns/synapse-interface
+
+
+
+
+
+## [0.1.178](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.1.177...@synapsecns/synapse-interface@0.1.178) (2023-11-09)
+
+**Note:** Version bump only for package @synapsecns/synapse-interface
+
+
+
+
+
+## [0.1.177](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.1.176...@synapsecns/synapse-interface@0.1.177) (2023-11-09)
+
+**Note:** Version bump only for package @synapsecns/synapse-interface
+
+
+
+
+
 ## [0.1.176](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.1.175...@synapsecns/synapse-interface@0.1.176) (2023-11-08)
 
 **Note:** Version bump only for package @synapsecns/synapse-interface